use a simplified / opt-in Component architecture
Previously, a `vecty.Composite` struct was embedded within user-defined
components such that they would not have to define the `Component.Node` method
or handle reconciliation of the body (rendered component). This meant you still
had to:

- Define an `Apply` method which would append your component as a child of an
  `Element` (boilerplate code).
- Define an `Reconcile` method regardless of whether or not your component
  needs to do anything _special_ during reconciliation (copying component
  state, for example).
- Wire up `Composite.RenderFunc` to point to your render function.

In the context of internal component types like `Element` and `textComponent`,
`Reconcile` is a very important part of code and contextually makes much sense.
Likewise, for markup types like `property`, `data`, `style`, etc. the `Apply`
method also makes a lot of sense because it is so important in these contexts.

However, for user components, without already having explained how Vecty/React
like libraries work internally like in the cases of `Element` or `property`,
it's hard to document and describe to newcomers why this boilerplate (`Apply`)
and wiring (`Composite.RenderFunc`) is needed, or why there is no default
`Reconcile` method.

If instead we define a `Core` struct with basic or 'default' implementations of
these methods (`Apply`, `Reconcile` and the `RenderFunc`) then we can put
ourselves in a position where we describe these methods as an opt-in 'feature'
instead of an up-front learning cost. This makes it easier to learn Vecty right
off the bat.

The only caveat is that it's hard to write a `Render` function at an embedded
`*vecty.Core` struct level, because there is no access to the parent struct,
the typical problem of trying to map a `super` OOP model to a non-OOP language
like Go. To address this, we do require that users write some boilerplate
around component creation by writing their own `NewMyComponent` function:

```
func NewMyComponent(props *MyComponentProps) *MyComponent {
	c := &MyComponent{
		MyComponentProps: props,
	}
	c.Core = vecty.New(c)
	return c
}
```

This way the `*vecty.Core` struct has a reference to the parent struct / 'super class'
if you will. Traditionally, this introduces a problem of garbage collection not
working due to cyclic references; but the good news is that the DOM
traditionally operates in this way anyway and thus all we need is a `Unmount`
handler to properly break this cyclic reference and allow GC to free the
object. This is very natural.