Add Component.Unmount for unmounting components.
Prior to this change, unmounting of components was done automatically within
`Element.Reconcile` because it is inherently part of removing elements that
were part of the previous render but not the current one (element removal).

After this change `Component` has a new constraint, an `Unmount` method which
makes it possible for each component to know when it is about to have its DOM
elements etc. removed. In particular, this is useful for user components that
might interact with the DOM (like adding a global event handler of some such),
and also valuable by `*Core` for breaking the Core <-> Component parent struct
/ 'superclass' link for proper (or perhaps just quicker) garbage collection.