event: start all func docstrings with the func name
This causes the generated documentation in the event package to always
follow the Go style convention that the docstring should start with the
function name.

This can be done reliably because the [documentation source](https://developer.mozilla.org/en-US/docs/Web/Events)
from which this documentation is generated is under the column "Fired when..."
and thus our docstrings will always read as correct paragraphs.