elem: start all func docstrings with the func name
This causes the generated documentation in the elem package to always
follow the Go style convention that the docstring should start with the
function name.

This is less relaible than the event package because the source documentation
is not laid out in such a trivial way. However, this still produces a
number of significantly better documented functions. Many more can be improved
by updating the MDN to use more consistent wording (right now it's quite
sporiadic).