use the same TextNode when restoring Text("") against Text("")
Checking that `text != ""` is not good enough because a text node _could_
contain no text (and there are valid use cases for such). So instead we
should rely on `tag == ""` to signify a text node.

This fixes an issue where restoration of an empty `Text("")` against
previous empty `Text("")` would create a new node instead of properly
being no-op.