Split `Restorer` skip functionality into `RenderSkipper` interface
- The `Restorer` interface is now just `Restore(Component)` (previously `Restore(Component) bool`).
- A `RenderSkipper` interface is added, where `SkipRender(Component) bool` replaces the old `Restore` return value behavior.

A notable change here is that the `prev` argument to `SkipRender` is a `Copy` of the `Component` at the time of the last `Render`, while the `prev` argument to `Restore` is not as it needs to copy component state (struct fields) from the old component instance to the new one.

See https://github.com/gopherjs/vecty/pull/112#issuecomment-302606385 for background.