implement persistent component lifecycle
Prior to this change a component's instance would be outright replaced
if the parent rendered a new component pointer; now instead of doing that,
we follow the react-like approach of keeping the old component pointer (the
one first rendered) as the persistent component instance. The parent
properties are copied over using reflection.

See https://github.com/gopherjs/vecty/issues/115#issuecomment-320491498 for further details.