handle nil *HTML in render (#138)
This change fixes a regression introduced by PR #134.

The basic issue is that when `nil` is encountered, it is treated correctly as a nil render,
but when `(*HTML)(nil)` is encountered, it is not. I discovered this while porting
Go-Package-Store (which uses Vecty) over to the new API.

The following code effectively demonstrates the issue:

```Go
package main

import (
        "github.com/gopherjs/vecty"
        "github.com/gopherjs/vecty/elem"
)

func main() {
        vecty.RenderBody(&PageView{})
}

type PageView struct {
        vecty.Core
}

func (p *PageView) Render() *vecty.HTML {
        return elem.Body(
                // this works:
                //nil,

                // this does not:
                p.returnsNil(),
        )
}

func (p *PageView) returnsNil() *vecty.HTML {
        return nil
}
```