renderer: Perform Rerender in batches, gated by requestAnimationFrame (#152)
By performing Rerender in batches, the main thread is blocked for
significantly less time in busy apps, leading to better responsiveness.

This change also deduplicates multiple Rerender calls for the same
Component in a single batch period. This can significantly improve
performance for busy apps, but carries the caveat that Components must
not rely on side-effects triggered by transitory states inside their
Render functions.

A time budget is applied to rendering each batch, in an attempt to
maintain a 60fps frame-rate.