hint at map allocation size when clearing index map of batchRenderer
In https://github.com/gopherjs/vecty/pull/152 I hinted that I had one
last suggestion, and this is it.

When we clear `b.idx`, we allocate a new map in its place. This map is
likely to have the same size as the old `b.idx`, so using this as an
allocation size hint in the call to `make` is a good idea.