correctly invoke Mount for components inside a List (#167)
I noticed that we accidently discard `pendingMounts` for any `Component` that
implements the `Mounter` interface inside a `List`. For example:

```
vecty.If(true, &SomeComponent{})
```

`SomeComponent.Mount` would never be invoked because it returns a `List` and
would have it's `pendingMounts` discarded. This fixes that, and I've verified
we do not discard `pendingMounts` by accident anywhere else.