example/markdown: fix building under GOARCH=wasm with GO111MODULE=off
d330d9e fixed the import path of blackfriday to reflect version 2, however
this import path is in fact not compatible with building in _both_ GO111MODULE=off
and GO111MODULE=on modes, which is something that is desirable to support for
now.

Using the gopkg.in import path temporarily works around this, and we can
revisit the issue after GO111MODULE's default changes to on or off.

Fixes CI.